+++ src/main/resources/db/migration/procedures/flow/workQueue/R__get_work_queue_metrics_for_events.sql
@@ -45,8 +45,8 @@ declare
 --   v_use_event_data                    boolean;
 --   v_smartlist_id                      bigint;
 --   v_color                             varchar;
-  v_long_window                       bigint;
-  v_short_window                      bigint;
+  v_long_window                       integer;
+  v_short_window                      integer;
   v_long_window_duration_type_id      bigint;
   v_short_window_duration_type_id     bigint;
   v_expected_cycle                    bigint;
